catalog: Fixed some problems caused by not parsin properly long ids.

+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/db/mongodb/converters/ProjectConverter.java
@@ -36,7 +36,7 @@ public class ProjectConverter extends GenericConverter<Project, Document> {
         Document document = null;
         try {
             document = Document.parse(projectWriter.writeValueAsString(object));
-            document.put("id", document.getLong("id"));
+            document.put("id", document.getInteger("id").longValue());
         } catch (JsonProcessingException e) {
             e.printStackTrace();
         }
